requirements.txt
numpy==1.26.3
opencv-python==4.10.0.82
tensorboard
tqdm
pyyaml
h5py
scipy
joblib
timm
fvcore
pyfftw
kornia==0.1.4
matplotlib
